+++ packages/testbot/features/convo.js
@@ -8,18 +8,16 @@ module.exports = function(controller) {
     welcome.say('Hey!');
     welcome.ask({
         text: ['Check this out...', 'YO! Check it.', 'HELLO HUME...'],
-        channelData: { 
-            quick_replies: [
-                {
-                    title: 'Foo',
-                    payload: 'foo',
-                },
-                {
-                    title: 'Bar',
-                    payload: 'bar',
-                }
-            ],
-        },
+        quick_replies: [
+            {
+                title: 'Foo',
+                payload: 'foo',
+            },
+            {
+                title: 'Bar',
+                payload: 'bar',
+            }
+        ],
     },async(answer, convo, bot) => {
         // noop.
     },{key: 'waterfall_results'});